pyproject.toml
[project]
name = "adk-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.14.0",
    "google-generativeai>=0.8.5",
    "litellm>=1.77.1",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uuid>=1.30",
    "yfinance>=0.2.65",
]
